values are no longer scaled in pips

+++ cli/config_file/signal_strat/parse_signal_strat_args.py
@@ -9,7 +9,7 @@ def parse_signal_strat_args(config_module) -> SimpleNamespace:
     ma_fnc = parse_arg(config_module.MA_FNC, '--ma_fnc', type=valid_ma_fnc, required=True)
     short_window = parse_arg(config_module.SHORT_WINDOW, '--short_window', type=int, required=True)
     long_window = parse_arg(config_module.LONG_WINDOW, '--long_window', type=int, required=True)
-    delta = parse_arg(config_module.DELTA, '--delta', type=int, required=True)
+    delta = parse_arg(config_module.DELTA, '--delta', type=float, required=True)
     confidence_window = parse_arg(config_module.CONFIDENCE_WINDOW, '--confidence_window', type=int, required=True)
     return SimpleNamespace(
         ma_fnc=ma_fnc,
